audioconvrec: new script for audio conversion
Script that handles flac/mp3/ogg conversion.

The old script audioconv works but requires plenty of manual file
management to work properly (transferring files to convert from
folder A to B, etc.)

This script is much better since it crawls into the conversion
directory and looks for files to convert. The output directory
is also made similar to the input directory, so the files are
still grouped after conversion.

Lot's of manual labor automated. Happy man. =)